requirements.txt
openai
sentence-transformers
numpy
pillow